+++ swim/macros/Personal Health Centre.js
@@ -8,7 +8,7 @@ function main() {
     }
     // Checking for SWADE Spices & Flavours and setting up the Benny image.
     let bennyImage = "icons/commodities/currency/coin-embossed-octopus-gold.webp";
-    if (game.modules.get("swade-spices").active) {
+    if (game.modules.get("swade-spices")?.active) {
         let benny_Back = game.settings.get(
             'swade-spices', 'bennyBack');
         if (benny_Back) {
@@ -457,7 +457,7 @@ for (let fatiguePotion of ownedFatiguePotions) {
         //Chat Message to let the everyone know a benny was spent
         ChatMessage.create({
             user: game.user._id,
-            content: `<p><img src="${bennyImage}"" width="25" height="25" /> ${game.user.name} spent a Benny for ${token.name}.</p>`,
+            content: `<p><img style="border: none;" src="${bennyImage}"" width="25" height="25" /> ${game.user.name} spent a Benny for ${token.name}.</p>`,
         });
     }
 
@@ -604,7 +604,7 @@ for (let fatiguePotion of ownedFatiguePotions) {
                             speaker: {
                                 alias: token.name
                             },
-                            content: `<img src="${potion_icon}" alt="" width="25" height="25" /> ${token.name} uses a ${selectedPotion} to heal ${genericHealWounds} wound(s).`
+                            content: `<img style="border: none;" src="${potion_icon}" alt="" width="25" height="25" /> ${token.name} uses a ${selectedPotion} to heal ${genericHealWounds} wound(s).`
                         })
                         if (potionSFX) {
                             let audioDuration = AudioHelper.play({ src: `${potionSFX}` }, true)._duration;
@@ -653,7 +653,7 @@ for (let fatiguePotion of ownedFatiguePotions) {
                             speaker: {
                                 alias: token.name
                             },
-                            content: `<img src="${potion_icon}" alt="" width="25" height="25" /> ${token.name} uses a ${selectedPotion} to cure ${genericHealFatigue} level(s) of Fatigue.`
+                            content: `<img style="border: none;" src="${potion_icon}" alt="" width="25" height="25" /> ${token.name} uses a ${selectedPotion} to cure ${genericHealFatigue} level(s) of Fatigue.`
                         })
                         if (potionSFX) {
                             let audioDuration = AudioHelper.play({ src: `${potionSFX}` }, true)._duration;
@@ -742,4 +742,4 @@ for (let fatiguePotion of ownedFatiguePotions) {
     }
 }
 
-// v.3.0.2 By SalieriC#8263; fixing bugs supported by FloRad#2142. Potion usage inspired by grendel111111#1603; asynchronous playback of sfx by Freeze#2689.
+// v.3.1.1 By SalieriC#8263; fixing bugs supported by FloRad#2142. Potion usage inspired by grendel111111#1603; asynchronous playback of sfx by Freeze#2689.