+++ swim/macros/Soak Damage.js
@@ -8,7 +8,7 @@ function main() {
     }
     // Checking for SWADE Spices & Flavours and setting up the Benny image.
     let bennyImage = "icons/commodities/currency/coin-embossed-octopus-gold.webp";
-    if (game.modules.get("swade-spices").active) {
+    if (game.modules.get("swade-spices")?.active) {
         let benny_Back = game.settings.get(
             'swade-spices', 'bennyBack');
         if (benny_Back) {
@@ -223,7 +223,7 @@ function main() {
         //Chat Message to let the everyone know a benny was spent
         ChatMessage.create({
             user: game.user._id,
-            content: `<p><img src="${bennyImage}"" width="25" height="25" /> ${game.user.name} spent a Benny for ${token.name}.</p>`,
+            content: `<p><img style="border: none;" src="${bennyImage}"" width="25" height="25" /> ${game.user.name} spent a Benny for ${token.name}.</p>`,
         });
     }
 
@@ -363,4 +363,4 @@ function main() {
     }
 }
 
-// V2.4.0 Code by SalieriC#8263. Critical Failure awareness by Kekilla#7036 Testing and bug-chasing: javierrivera#4813.
+// V2.5.1 Code by SalieriC#8263. Critical Failure awareness by Kekilla#7036 Testing and bug-chasing: javierrivera#4813.