+++ swim/macros/Spend Benny.js
@@ -13,7 +13,7 @@ async function main() {
 
     // Checking for SWADE Spices & Flavours and setting up the Benny image.
     let bennyImage = "icons/commodities/currency/coin-embossed-octopus-gold.webp";
-    if (game.modules.get("swade-spices").active) {
+    if (game.modules.get("swade-spices")?.active) {
         let benny_Back = game.settings.get(
             'swade-spices', 'bennyBack');
         if (benny_Back) {
@@ -46,7 +46,7 @@ async function main() {
     //Chat Message to let the everyone know a benny was spent
     ChatMessage.create({
         user: game.user._id,
-        content: `<p><img src="${bennyImage}"" width="25" height="25" /> ${game.user.name} spent a Benny for ${token.name}.</p>`,
+        content: `<p><img style="border: none;" src="${bennyImage}"" width="25" height="25" /> ${game.user.name} spent a Benny for ${token.name}.</p>`,
     });
 }