+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1]
+### You need to update these macros:
+- Spend Benny
+- (Un-)Shake (both versions)
+- (Un-)Stun
+- Personal Health Centre
+- Power Point Management
+- Soak Damage
+### Fixed
+- Removed the border of images in chat messages.
+- Fixed a breaking bug that caused the module to not work if SWADE Spices & Flavours was not installed (thx @Khitan#9089 for bringing it to my attention).
+
 ## [0.6.0] - 2021-03-10
 ### You need to update these macros:
 - (Un-)Stunned